% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fq_bases}
\alias{fq_bases}
\title{FASTQ number of bases}
\usage{
fq_bases(infile)
}
\arguments{
\item{infile}{Path to fastq/fastq.gz file}
}
\value{
Numeric
}
\description{
Obtain the number of bases in a fastq/fastq.gz file
}
