% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fq_kmers}
\alias{fq_kmers}
\title{FASTQ k-mer counts}
\usage{
fq_kmers(infile, k, nth)
}
\arguments{
\item{infile}{Path to fastq/fastq.gz file}

\item{k}{kmer length}

\item{nth}{calculate kmer for every nth record - used to subsample big fastq files}
}
\value{
Numeric vector of k-mer counts (sorted by k-mer key)
}
\description{
Get k-mer counts in a fastq/fastq.gz file
}
