% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fq_reads}
\alias{fq_reads}
\title{FASTQ number of reads}
\usage{
fq_reads(infile)
}
\arguments{
\item{infile}{Path to fastq/fastq.gz file}
}
\value{
Numeric
}
\description{
Obtain the number of reads in a fastq/fastq.gz file
}
