% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fq_report.R
\name{fq_report}
\alias{fq_report}
\title{fastq html summary report}
\usage{
fq_report(inpath)
}
\arguments{
\item{inpath}{A path to folder containing fastq/fastq.gz files}
}
\description{
Generate html summary report for the fastq files in a folder
}
