% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fq_summary.R
\name{fq_summary}
\alias{fq_summary}
\title{FASTQ file summary}
\usage{
fq_summary(infile)
}
\arguments{
\item{infile}{A fastq/fastq.gz file}
}
\value{
A named list
}
\description{
Obtain summary statistics for the records of a fastq/fastq.gz file
}
