% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfaster2-package.R
\docType{package}
\name{rfaster2-package}
\alias{rfaster2}
\alias{rfaster2-package}
\title{rfaster2: Port Faster2 To R}
\description{
This package exports some of the functions of the Rust program faster2 in R. It can be used to get statistics about the records in a fastq/fastq.gz file.
}
\author{
\strong{Maintainer}: Angel Angelov \email{aangeloo@gmail.com}

}
\keyword{internal}
