% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_summary}
\alias{rust_summary}
\title{Return vector with summary values for the fastq file, this is not exported in the R package namespace}
\usage{
rust_summary(infile)
}
\description{
Return vector with summary values for the fastq file, this is not exported in the R package namespace
}
