% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fq_gc}
\alias{fq_gc}
\title{FASTQ GC contents}
\usage{
fq_gc(infile)
}
\arguments{
\item{infile}{Path to fastq/fastq.gz file}
}
\value{
Numeric vector with GC content (fraction) per record
}
\description{
Obtain per read GC contents from a fastq/fastq.gz file
}
