% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fq_quals}
\alias{fq_quals}
\title{FASTQ 'mean' quality scores}
\usage{
fq_quals(infile, phred, nth)
}
\arguments{
\item{infile}{Path to fastq/fastq.gz file}

\item{phred}{Logical, report Phred score (error probability otherwise)}

\item{nth}{calculate qual for every nth record - used to subsample big fastq files}
}
\value{
Numeric vector with 'mean' quality score per read
}
\description{
Obtain'mean' fastq read quality values. Mean Phred scores are calculated by converting to probabiliy, calculating mean quality, then converting back to Phred.
}
